Begin to setup flask. TODO: fix npm start error.
Renamed the default frontend folder
src -> `interface`. Flask server will be in `server`
The error I'm getting is that it can't find `index.js` in `src` where it is looking.
It might be because this local repo is in my dropbox. I will try moving it and seeing if that helps.