Add dependency instructions. Todo: Reroute src.
Still getting the same error about unable to do `npm start` because it is looking for index.js in `src`.

This is still happening in my non-dropbox clone. So, dropbox wasn't the reason for the problem.

We need to figure out what argument to add to `npm start` or package file to change to change where it is looking for the frontend code.